+++ packages/mad-tag-ocr/src/index.ts
@@ -1,0 +1,2 @@
+export { OCRCamera } from "./components/Camera/OCRCamera";
+export * from "./types";